+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased
 * Add AuthorizationAdjustment class and `authorizationAdjustments` to Transaction
-* Coinbase is no longer a supported payment method. Error code `93117` will be returned for Coinbase operations.
+* Coinbase is no longer a supported payment method. `PaymentMethodNoLongerSupported` will be returned for Coinbase operations.
 
 ## 2.2.0
 * Fixes wrap-promise issue (closes #102) - Thanks @targunp