+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Coinbase is no longer a supported payment method. `PaymentMethodNoLongerSupported` will be returned for Coinbase operations.
 * Add facilitated details to Transaction if present
 * Add `submit_for_settlement` to `Subscription.retryCharge`
+* Add `options` -> `paypal` -> `description` for creating and updating subscriptions
 
 ## 2.2.0
 * Fixes wrap-promise issue (closes #102) - Thanks @targunp